Fixed all the issues
Fixed memory leaks by using unique ptrs, a bit unwise to use smart
pointers if I opt for a single ownership system.

Also added comptrs, was an issue when getting a reference of a pointer.
By creating a contemporary pointer it takes the adress of this new
pointer instead of the original one. Uses get address of comptr